refactor(organize): extract shared tags, timeout, and scanner utilities
Consolidate normalizeTags/addTag into tags.ts, withTimeout into
timeout.ts, and reuse bookmarkNoteFiles in link-check by dropping its
private scanner copy. Also narrow the Wayback JSON parse to remove
unsafe-any lint errors. Behavior-preserving: tsc, build, and lint
(0 errors) pass.
