+++ src/timeout.ts
@@ -1,0 +1,17 @@
+/** Reject with "timeout" if `promise` does not settle within `ms`; otherwise pass
+ * its result through. Clears the timer on either outcome. */
+export function withTimeout<T>(promise: Promise<T>, ms: number): Promise<T> {
+	return new Promise<T>((resolve, reject) => {
+		const timer = window.setTimeout(() => reject(new Error("timeout")), ms);
+		promise.then(
+			(value) => {
+				window.clearTimeout(timer);
+				resolve(value);
+			},
+			(error) => {
+				window.clearTimeout(timer);
+				reject(error instanceof Error ? error : new Error(String(error)));
+			},
+		);
+	});
+}
